requirements.txt
dj-database-url==0.3.0

Django==1.8.1
gunicorn==19.3.0
psycopg2==2.6


SQLAlchemy==1.0.4

whitenoise==1.0.6
django-postgrespool==0.3.0
django-haystack==2.4.0
whoosh==2.7.0